+++ src/pglib/shared.jl
@@ -7,7 +7,9 @@ function constraint_thermal_limit_from(pm::PMs.AbstractPowerModel, n::Int, i::In
     t_bus = branch["t_bus"]
     f_idx = (i, f_bus, t_bus)
 
-    PMs.constraint_thermal_limit_from(pm, n, f_idx, branch["rate_a"]*scale)
+    if haskey(branch, "rate_a")
+        PMs.constraint_thermal_limit_from(pm, n, f_idx, branch["rate_a"]*scale)
+    end
 end
 constraint_thermal_limit_from(pm::PMs.AbstractPowerModel, i::Int; scale = 1.0) = constraint_thermal_limit_from(pm, pm.cnw, i; scale=scale)
 
@@ -18,7 +20,9 @@ function constraint_thermal_limit_to(pm::PMs.AbstractPowerModel, n::Int, i::Int;
     t_bus = branch["t_bus"]
     t_idx = (i, t_bus, f_bus)
 
-    PMs.constraint_thermal_limit_to(pm, n, t_idx, branch["rate_a"]*scale)
+    if haskey(branch, "rate_a")
+        PMs.constraint_thermal_limit_to(pm, n, t_idx, branch["rate_a"]*scale)
+    end
 end
 constraint_thermal_limit_to(pm::PMs.AbstractPowerModel, i::Int; scale = 1.0) = constraint_thermal_limit_to(pm, pm.cnw, i; scale=scale)